# openapi.yml
openapi: "3.0.2"
info:
  title: Comfy API
  version: "1.0"
servers:
  - url: https://api.comfy.org
paths:
...
  /upload-artifact:
    post:
      summary: Receive artifacts (output files) from the ComfyUI GitHub Action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repo:
                  type: string
                  description: Repository name
                job_id:
                  type: string
                  description: Unique identifier for the job
                run_id:
                  type: string
                  description: Unique identifier for the run
                os:
                  type: string
                  description: Operating system used in the run
                cuda_version:
                  type: string
                  description: Cuda version.
                bucket_name:
                  type: string
                  description: The name of the bucket where the output files are stored
                output_files_gcs_paths:
                  type: string
                  description: A comma separated string that contains GCS path(s) to output files. eg. gs://bucket-name/output, gs://bucket-name/output2
                comfy_logs_gcs_path:
                  type: string
                  description: The path to ComfyUI logs. eg. gs://bucket-name/logs
                commit_hash:
                  type: string
                commit_time:
                  type: string
                  description: The time of the commit in the format of "YYYY-MM-DDTHH:MM:SSZ" (2016-10-10T00:00:00Z)
                commit_message:
                  type: string
                  description: The commit message
                workflow_name:
                  type: string
                  description: The name of the workflow
                branch_name:
                  type: string
                start_time:
                  type: integer
                  format: int64
                  description: The start time of the job as a Unix timestamp.
                end_time:
                  type: integer
                  format: int64
                  description: The end time of the job as a Unix timestamp.
              required:
                - repo
                - job_id
                - run_id
                - os
                - gpu_type
                - commit_hash
                - commit_time
                - commit_message
                - branch_name
                - workflow_name
                - start_time
                - end_time

      responses:
        '200':
          description: Successfully received the artifact details
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid request
        '500':
          description: Internal server error
  /gitcommit:
    get:
      summary: Retrieve CI data for a given commit
      description: Returns all runs, jobs, job results, and storage files associated with a given commit.
      parameters:
        - in: query
          name: commitId
          required: false
          schema:
            type: string
          description: The ID of the commit to fetch data for.
        - in: query
          name: operatingSystem
          required: false
          schema:
            type: string
          description: The operating system to filter the CI data by.
        - in: query
          name: workflowName
          required: false
          schema:
            type: string
          description: The name of the workflow to filter the CI data by.
        - in: query
          name: branch
          required: false
          schema:
            type: string
          description: The branch of the gitcommit to filter the CI data by.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
          description: The page number to retrieve.
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
            default: 10
          description: The number of items to include per page.
        - in: query
          name: repoName
          required: false
          schema:
            type: string
            default: comfyanonymous/ComfyUI
          description: The repo to filter by.
      responses:
        '200':
          description: An object containing runs, jobs, job results, and storage files
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobResults:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActionJobResult'
                  totalNumberOfPages:
                    type: integer
        '404':
          description: Commit not found
        '500':
          description: Internal server error
  /branch:
    get:
      summary: Retrieve all distinct branches for a given repo
      description: Returns all branches for a given repo.
      parameters:
        - in: query
          name: repo_name
          required: true
          schema:
            type: string
            default: comfyanonymous/ComfyUI
          description: The repo to filter by.
      responses:
        '200':
          description: An array of branches
          content:
            application/json:
              schema:
                type: object
                properties:
                  branches:
                    type: array
                    items:
                      type: string
        '404':
          description: Repo not found
        '500':
          description: Internal server error
...
  /nodes:
    get:
      summary: Retrieves a list of nodes
      description: Returns a paginated list of nodes across all publishers.
      operationId: listAllNodes
      tags:
        - Nodes
      parameters:
        - in: query
          name: page
          description: Page number of the nodes list
          required: false
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          description: Number of nodes to return per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A paginated list of nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Total number of nodes available
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'
                  page:
                    type: integer
                    description: Current page number
                  limit:
                    type: integer
                    description: Maximum number of nodes per page
                  totalPages:
                    type: integer
                    description: Total number of pages available
        '400':
          description: Invalid input, object invalid
        '404':
          description: Not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    PublisherUser:
      type: object
      properties:
        id:
          type: string
          description: The unique id for this user.
        email:
          type: string
          description: The email address for this user.
        name:
          type: string
          description: The name for this user.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    ActionJobResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the job result
        workflow_name:
          type: string
          description: Name of the workflow
        operating_system:
          type: string
          description: Operating system used
        gpu_type:
          type: string
          description: GPU type used
        pytorch_version:
          type: string
          description: PyTorch version used
        action_run_id:
          type: string
          description: Identifier of the run this result belongs to
        commit_hash:
          type: string
          description: The hash of the commit
        commit_id:
          type: string
          description: The ID of the commit
        commit_time:
          type: integer
          format: int64
          description: The Unix timestamp when the commit was made
        commit_message:
          type: string
          description: The message of the commit
        git_repo:
          type: string
          description: The repository name
        start_time:
          type: integer
          format: int64
          description: The start time of the job as a Unix timestamp.
        end_time:
          type: integer
          format: int64
          description: The end time of the job as a Unix timestamp.
        storage_file:
          $ref: "#/components/schemas/StorageFile"
    StorageFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the storage file
        file_path:
          type: string
          description: Path to the file in storage
        public_url:
          type: string
          description: Public URL
    Publisher:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
          description: The unique identifier for the publisher. It's akin to a username. Should be lowercase.
        description:
          type: string
        website:
          type: string
        support:
          type: string
        source_code_repo:
          type: string
        logo:
          type: string
          description: URL to the publisher's logo.
        createdAt:
          type: string
          format: date-time
          description: The date and time the publisher was created.
        members:
          type: array
          items:
            $ref: "#/components/schemas/PublisherMember"
          description: A list of members in the publisher.
    PublisherMember:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the publisher member.
        user:
          $ref: "#/components/schemas/PublisherUser"
          description: The user associated with this publisher member.
        role:
          type: string
          description: The role of the user in the publisher.
    Node:
      type: object
      properties:
        id:
          type: string
          description: "The unique identifier of the node."
        name:
          type: string
          description: The display name of the node.
        description:
          type: string
        author:
          type: string
        license:
          type: string
          description: The path to the LICENSE file in the node's repository.
        icon:
          type: string
          description: URL to the node's icon.
        repository:
          type: string
          description: URL to the node's repository.
        tags:
          type: array
          items:
            type: string
        latest_version:
          $ref: "#/components/schemas/NodeVersion"
          description: The latest version of the node.
        rating:
          type: number
          description: The average rating of the node.
        downloads:
          type: integer
          description: The number of downloads of the node.
        publisher:
          $ref: "#/components/schemas/Publisher"
          description: The publisher of the node.
    NodeVersion:
      type: object
      properties:
        id:
          type: string
        version:
          type: string
          description: The version identifier, following semantic versioning. Must be unique for the node.
        createdAt:
          type: string
          format: date-time
          description: The date and time the version was created.
        changelog:
          type: string
          description: Summary of changes made in this version
        dependencies:
          type: array
          items:
            type: string
          description: A list of pip dependencies required by the node.
        downloadUrl:
          type: string
          description: "[Output Only] URL to download this version of the node"
        deprecated:
          type: boolean
          description: Indicates if this version is deprecated.
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT